Add example deliver and train scripts
These are based on the wrapper scripts I use to file my own inbound mail.
They expect to live alongside their dictionaries in a 'zf' subdirectory
at the top level of a maildir.

Dictionaries are trained on whole messages and message headers from the
last 800 days of email in the Read and Junk folders respectively.

On any failure, the deliver script exits with status 75 which should
avoid a bounce and temporarily defer the pipe delivery in most MTAs.