Add an optional header helper
This extracts specified header values from messages, optionally parsing
dates into unix timestamps and address lists into individual addresses.

It is written in Go to take advantage of built-in RFC 5322 support in the
standard library and to avoid heavy-duty string wrangling of untrusted
message text in C. It isn't necessary for using the rest of zf, so don't
try to build or install it if a Go compiler is not available.